may/requirements.txt
accelerate==0.21.0
aiofile==3.8.1
anyio==3.6.2
appnope==0.1.4
arrow==1.2.3
asttokens==3.0.0
attrs==22.2.0
baidu-aip==4.16.9
beautifulsoup4==4.11.2
bs4==0.0.1
caio==0.9.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
circular==0.1
click==8.1.6
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coverage==7.0.5
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.1.0
executing==2.2.0
file-io==0.3.9
filelock==3.12.2
fonttools==4.57.0
fsspec==2022.11.0
gy-multiprocessing==0.2.4
huggingface-hub==0.16.4
idna==3.4
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazycls==0.0.1
legacy==0.1.6
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==9.3.0
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
openpyxl==3.1.5
packaging==23.0
pandas==1.5.3
parso==0.8.4
peewee==3.15.4
peft==0.4.0
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.3.7
plotly==6.1.1
pluggy==1.0.0
prompt_toolkit==3.0.50
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
PyEmail==0.0.1
pyfiglet==0.8.post1
pygame==2.3.0
pygame-utils==0.3.0
Pygments==2.19.1
PyMySQL==1.0.2
pyparsing==3.2.3
pytest==7.2.1
pytest-cov==2.12.1
python-dateutil==2.8.2
python-docx==1.1.2
python-time==0.3.0
pytz==2021.3
PyYAML==6.0.1
pyzmq==26.3.0
rand==0.3.0
regex==2023.6.3
requests==2.28.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4.2
tqdm==4.65.0
traitlets==5.14.3
transformers==4.31.0
tushare==1.2.89
typing_extensions==4.13.2
urllib3==1.26.14
wcwidth==0.2.13
websocket-client==0.57.0
